\name{P}
\alias{P}
\docType{data}
\title{
P-values
}
\description{
Each p-value (i.e., each row) corresponds to one participant in the study.
}
\usage{data(P)}
\format{
  A data frame with 7 observations on the following variable.
  \describe{
    \item{\code{V1}}{P-value}
  }
}
\examples{
data(P)
}
\keyword{datasets}
