\name{SCMA-package}
\alias{SCMA-package}
\docType{package}
\title{
Single-Case Meta-Analysis
}
\description{
Perform meta-analysis of single-case experiments, including calculating various effect size measures (SMD, PND and PEM) and probability combining (additive and multiplicative method).
}
\details{
\tabular{ll}{
Package: \tab SCMA\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-01-13\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte and Patrick Onghena

Maintainer: Tamal Kumar De <tamalkumar.de@kuleuven.be>
}

\keyword{ package }
