\name{SCMA-package}
\alias{SCMA-package}
\docType{package}
\title{
Single-Case Meta-Analysis
}
\description{
A package for probability combining (additive and multiplicative method) and for calculating some effect size measures on single-case data (SMD, PND and PEM). 
}
\details{
\tabular{ll}{
Package: \tab SCMA\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-01-13\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte and Patrick Onghena

Maintainer: Bart Michiels <bart.michiels@ppw.kuleuven.be>
}

\keyword{ package }
